feat : Get Department Details by Id Feature Done
